compositor: guard scene/compose math against hostile geometry (6 overflow fixes)
Hostile-geometry sweep of OUR scene/compose/present math (the layer the
protocol-level hostile-client sweep #173 didn't reach - smithay guards its
protocol handling, but our pixel/geometry math had the same attacker-size
-> unchecked add/mul -> panic/OOB/multi-GiB pattern the driver sweeps found).
19-test hostile_scene.rs; six real bugs found + fixed (each FAILED pre-fix):

- resample_nearest panic on a zero/negative-dim source: i32::clamp(min,max)
  with min>max panics. Refuse the frame (present.rs).
- Unbounded/overflow alloc: a client wp_viewport.set_destination of a few
  billion flows through compose -> logical_size -> vec![0u8; dw*dh*4];
  dw*dh*4 overflows i32 (debug panic) / multi-GiB alloc. present() gate
  computes real output dims per viewport/transform arm and reports
  offscreen (skip, no pacing advance) when dst axis <=0 or > MAX_PRESENT_DIM
  (16384) or a buffer is inconsistent with its rgba.
- transform_buffer overflow + OOB slice: (ow*oh*4) as i32 overflow + index
  math overflow + buf.rgba[si..si+4] OOB if rgba short. Fixed by the gate +
  all index math in usize.
- damage.rs Rect right()/bottom()/translate()/union() i32 overflow with
  hostile subsurface offset+size corrupted containment/occlusion. saturating.
- scene.rs offset accumulation (collect_subtree_offsets, popup_offset_to_
  toplevel) overflow across a chain of absurd set_position/popup geometries.
  saturating_add.
- read_shm_rgba (re-checked per #173): stride < w*4 and vec![0u8; w*h*4]
  compute w*4 / w*h*4 in i32 before rejecting geometry; hostile width near
  i32::MAX panics. Widened to i64/usize.

default: 103 passed (94 + 9). smithay-adapter: 164 passed (145 + 19, 10
feature-gated present-path rasterizer cases).

+++ hl_wip-compositor/src/adapter/smithay/state.rs
@@ -1922,12 +1922,15 @@ fn buffer_transform_to_wl(t: BufferTransform) -> Transform {
 fn read_shm_rgba(buffer: &WlBuffer) -> Option<(StoredBuffer, Format)> {
     let result = with_buffer_contents(buffer, |ptr, len, data| {
         let (w, h, stride, offset) = (data.width, data.height, data.stride, data.offset);
-        if w <= 0 || h <= 0 || stride < w * 4 || offset < 0 {
+        // `w * 4` is computed in `i64` so a hostile width near `i32::MAX` (reachable with a large `wl_shm`
+        // pool) cannot overflow the row-stride check itself before the geometry is rejected.
+        if w <= 0 || h <= 0 || (stride as i64) < w as i64 * 4 || offset < 0 {
             return None;
         }
-        // Highest byte read = offset + (h-1)*stride + w*4; must fit the mapping.
+        // Highest byte read = offset + (h-1)*stride + w*4; must fit the mapping. All widened to `usize`
+        // (each factor is now known positive) so the bound check can never overflow before it fires.
         let last_row = offset as usize + (h as usize - 1) * stride as usize;
-        if last_row.checked_add((w * 4) as usize).map(|m| m > len).unwrap_or(true) {
+        if last_row.checked_add(w as usize * 4).map(|m| m > len).unwrap_or(true) {
             return None;
         }
         // `format` is the neutral opaque/alpha distinction (drives blend); `swap_rb` selects channel
@@ -1939,7 +1942,9 @@ fn read_shm_rgba(buffer: &WlBuffer) -> Option<(StoredBuffer, Format)> {
             // Argb8888 and any other advertised/unknown format fall through to ARGB semantics.
             _ => (Format::Argb8888, false, true),
         };
-        let mut rgba = vec![0u8; (w * h * 4) as usize];
+        // `w`/`h` are positive and `w·h·4 <= len` (the bound check above), so this `usize` product is
+        // bounded by the mapping size and cannot overflow.
+        let mut rgba = vec![0u8; w as usize * h as usize * 4];
         for y in 0..h {
             let row = offset as isize + y as isize * stride as isize;
             for x in 0..w {
